requirements.txt
fastapi
uvicorn[standard]
groq
python-dotenv
supabase
tiktoken
sentence-transformers
